requirements.txt
pycairo
PyGObject
matplotlib
shapely
descartes
svgwrite
ipython
numpy
nptyping
build
coverage
sphinx
python-docs-theme
pyright
jupyter
nbconvert
sconstool-nbconvert
PyPDF2
pytest
pynvim
tox
debugpy
schemdraw
sympy
nbformat
rich
Pillow
matplotlib-backend-kitty
opencv-python
